Convert pdf to images if still present

+++ dataset/docile.py
@@ -11,19 +11,22 @@ class DocILE(Dataset):
         split: str
     ) -> None:
         super().__init__(task=task, split=split)
+        self._convert_pdf_to_img()
         self._load_data()
 
     def _convert_pdf_to_img(self):
-        split_file: list[str] = json.load(open(f"./data/docile/{self.split}.json", "r"))
-
-        for fn in split_file: 
-            images = convert_from_path(
-                pdf_path=f"./data/docile/pdfs/{fn}.pdf",
-                dpi=200,
-                fmt="jpg"
-            )
-            images[0].save(f"./data/docile/pdfs/{fn}.jpg")
-            os.remove(f"./data/docile/pdfs/{fn}.pdf")
+        '''
+        Converts pdfs into images
+        '''
+        for fn in os.listdir(f"./data/docile/pdfs"): 
+            if not fn.endswith(".pdf"):
+                images = convert_from_path(
+                    pdf_path=f"./data/docile/pdfs/{fn}.pdf",
+                    dpi=200,
+                    fmt="jpg"
+                )
+                images[0].save(f"./data/docile/pdfs/{fn}.jpg")
+                os.remove(f"./data/docile/pdfs/{fn}.pdf")
 
     def _load_data(self) -> None:
         split_file: list[str] = json.load(open(f"./data/docile/{self.split}.json", "r"))
